Give the testing extension the -blocking suffix every other paired leaf has
occurrent-testing-junit-jupiter is blocking-only and said nothing about it,
unlike occurrent-subscription-api-blocking, occurrent-eventstore-api-blocking
and occurrent-tck-eventstore-blocking. A reactive twin is coming in 0.32.0
(#530), so the pair needs the suffix.

Doing it now because the artifact is unreleased, its changelog entry still
living under the next-version heading, so per the release-status rule in
AGENTS.md it renames with no migration path. After 0.32.0 the same one-word
change would need an UpgradeToOccurrent recipe and a migration note.

The Java and Kotlin packages take .blocking too, so the twin's extension can
keep the same simple name without two of them colliding on one classpath.

ADR 82 gets a note rather than a rewritten table: the table records the layout
as decided then, and only the leaf's name changed.

Resolves #529.
