+++ changelog.md
@@ -20,7 +20,7 @@
 * A range filter on a numeric CloudEvent extension no longer fails against the in-memory store. Comparing a value to an operand of a different numeric type, as in `Filter.filter("streamversion", gt(5))` where the stored value is a `Long`, threw `ClassCastException` instead of comparing the two. Numbers are now compared by value regardless of their Java type, and a comparison between genuinely different types, such as a number and a string, matches nothing rather than failing, which is what MongoDB does.
 * `SpringMongoSubscriptionModelConfig.autoStartup(false)` gives you a subscription model that is created stopped, so a subscription registered on it is paused from the outset and no change stream opens until you call `start()` or `resumeSubscription(id)`. Use it to bring subscriptions up under your own control, behind a leader election or a health check, or in a test that picks which ones run. It also decides what `isAutoStartup()` reports, so a model you register as a Spring bean is not started for you either, which it previously was regardless of how you configured it. The default is unchanged.
 * The synchronous and push subscription models can now be stopped, started and paused like every other subscription model, on both the blocking and reactive stacks. Before this they had no life cycle at all, so a test that stopped every subscription still had its synchronous projections running, and `startAll()` skipped them. Note what stopping means for these models: an event handed to a stopped model, or to a paused subscription, is dropped rather than held, and resuming does not replay it. That matters most for a synchronous projection, since the write still succeeds while the projection does not run. `shutdown()` is the one that is not reversible: it drops every registration and releases the ids, so a shut-down model delivers nothing even after `start()`. [ADR 85](doc/architecture/decisions/0085-every-subscription-model-can-be-stopped.md) explains why this is the right trade and why the previous decision not to have a life cycle did not hold.
-* A new `occurrent-testing-junit-jupiter` artifact gives you a JUnit 5 extension that keeps every subscription stopped while a test runs, so a test only runs the subscriptions it names. `OccurrentSubscriptionsExtension.stoppedByDefault(subscriptionModel)` stops them before and after each test, `start("someId")` resumes one and blocks until it has actually started so you cannot forget to wait, `alwaysStart("someId")` resumes a fixed set for a whole test class, and `startAll()` starts every subscription the model has, for the one test that needs to see them working together. Naming an id that does not exist fails with the ids that do. It depends on nothing but JUnit and the blocking subscription API, so it works without Spring and without a container. The point of the default is that adding a subscription to your application can no longer change what an existing test asserts. [ADR 82](doc/architecture/decisions/0082-a-published-testing-module-for-application-authors.md) explains why this is published to Maven Central while `test-support` stays internal.
+* A new `occurrent-testing-junit-jupiter-blocking` artifact gives you a JUnit 5 extension that keeps every subscription stopped while a test runs, so a test only runs the subscriptions it names. `OccurrentSubscriptionsExtension.stoppedByDefault(subscriptionModel)` stops them before and after each test, `start("someId")` resumes one and blocks until it has actually started so you cannot forget to wait, `alwaysStart("someId")` resumes a fixed set for a whole test class, and `startAll()` starts every subscription the model has, for the one test that needs to see them working together. Naming an id that does not exist fails with the ids that do. It depends on nothing but JUnit and the blocking subscription API, so it works without Spring and without a container. The point of the default is that adding a subscription to your application can no longer change what an existing test asserts. [ADR 82](doc/architecture/decisions/0082-a-published-testing-module-for-application-authors.md) explains why this is published to Maven Central while `test-support` stays internal.
 * `SubscriptionModelLifeCycle.stop()` now documents that it leaves every running subscription paused, so you can resume one on its own with `resumeSubscription(id)` afterwards without starting the rest. Every subscription model already behaved this way, but nothing said so, which made it unsafe to build on.
 * A subscription model can now tell you which subscriptions it has. `IntrospectableSubscriptionModel.subscriptionIds()` returns every id it knows, running or paused, and `IntrospectableSubscriptionModel.of(subscriptionModel)` finds it behind any number of wrapping models so you do not have to unwrap by hand. The in-memory, Spring MongoDB and native MongoDB models implement it, and the competing consumer model reports its own consumers too, including one still waiting for the lock that its delegate has not been told about yet. Before this, `isRunning(id)` and `isPaused(id)` could answer for an id you already had, and nothing could list them. [ADR 83](doc/architecture/decisions/0083-a-subscription-model-reports-the-subscriptions-it-knows.md) explains why this is a separate capability rather than a method on `SubscriptionModelLifeCycle`.
 * `CatchupSubscriptionModel.stop()` now stops a catch-up replay that is already in flight. Before this it only stopped the live subscription model behind it, so a replay kept reading history and delivering events after `stop()` returned. `isRunning()`, `isRunning(String)` and `isPaused(String)` now also account for a replay in progress, which the live model cannot report because the subscription is not registered with it until the replay hands over.
